scraper/requirements.txt
# Core DB + env
psycopg[binary,pool]==3.2.10
python-dotenv==1.0.1

# JobSpy (Indeed, Google Jobs, ZipRecruiter, Glassdoor, LinkedIn). Requires Python 3.10+.
python-jobspy==1.1.82

# HTTP + parsing
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0
